Decode the product id with the same unsigned LCG
FS_SetRestrictions decodes the full game's productid.txt with Q_rand's
69069 LCG, also in signed int, so the multiply overflows from the
second byte on. It runs natively at every filesystem startup and
FS_Restart whenever productid.txt is present, which it is with the full
game's data.

Use the same unsigned multiply and add as Q_rand. The keystream, and so
the decoded 152-byte text, stays byte-identical.

tests/run_fs_product_id_tests.sh drives the real FS_SetRestrictions
through the fs_zip fixture. A pk3 holding the full game's productid.txt
must lift the restrictions, and copies with the first or last byte
changed must still be fatal. It also checks that the text is
fs_scrambledProductId decoded with the LCG in uint32_t. It builds with
-fsanitize=signed-integer-overflow -fno-sanitize-recover=all, at -O0
and -O2. The fs_config_mutable fixture no longer claims that the gate
overflows.

No other 69069 or similar signed LCG runs natively: bg_lib.c's rand()
is QVM-only, and net_chan.c's scramble is unsigned and compiled out.

Refs #426

+++ code/qcommon/files.c
@@ -3378,7 +3378,7 @@ static void FS_SetRestrictions( void ) {
 				if ( ( fs_scrambledProductId[i] ^ (seed&255) ) != productId[i] ) {
 					break;
 				}
-				seed = (69069 * seed + 1);
+				seed = (int)(69069U * (unsigned)seed + 1U);	// Q_rand's wrap, no signed overflow
 			}
 
 			FS_FreeFile( productId );
